requirements.txt
# tkinter should be installed via apt
pillow
platypus-Opt
matplotlib
pandas
seaborn